Implemented get albums of an artist.
Updated controllers, axios calls and types to make them compatible with the Spotify API.
To get the complete Albums, we have to first get the artist id, then make another call to get the artist's album's ids, and finally get those complete albums.
The returned albums to the Client-Side are sorted by popularity.